+++ gsrs-module-substances-core/src/main/java/gsrs/module/substance/services/CodeEntityService.java
@@ -20,7 +20,7 @@ import ix.core.validator.ValidatorCallback;
 import ix.ginas.models.v1.Code;
 import ix.ginas.models.v1.Reference;
 import ix.ginas.models.v1.Substance;
-import ix.ginas.utils.GinasProcessingStrategy;
+import ix.ginas.utils.validation.strategy.GinasProcessingStrategy;
 import ix.ginas.utils.validation.strategy.AcceptApplyAllProcessingStrategy;
 import ix.utils.Util;
 import org.springframework.beans.factory.annotation.Autowired;
