+++ gsrs-module-substances-core/src/main/java/ix/ginas/utils/validation/strategy/AcceptAndApplyAllWarningsProcessingStrategy.java
@@ -3,7 +3,6 @@ package ix.ginas.utils.validation.strategy;
 import gsrs.services.GroupService;
 import ix.core.validator.GinasProcessingMessage;
 import ix.core.validator.ValidationResponse;
-import ix.ginas.utils.GinasProcessingStrategy;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
